requirements.txt
amqp==5.1.0
anyjson==0.3.3
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==21.2.0
Authlib==0.15.3
backcall==0.2.0
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
bleach==3.3.1
celery==5.2.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
Django==3.1.7
django-celery==3.3.1
django-celery-beat==2.2.1
django-celery-results==2.3.0
django-cors-headers==3.11.0
django-crispy-forms==1.11.2
django-group-by==0.3.1
django-multiselectfield==0.1.12
django-rest==0.8.7
django-timedeltatemplatefilter==0.1.2
django-timezone-field==4.2.3
django-utf8field==1.0.0
djangorestframework==3.12.4
entrypoints==0.3
filelock==3.0.12
idna==2.10
install==1.3.4
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonify==0.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
kombu==5.2.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mysqlclient==2.0.3
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.1
oauthlib==3.1.0
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.1.2
platformdirs==2.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2==2.9.3
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2022.1
pywin32==301
pywinpty==1.1.3
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
requests==2.25.1
schedule==1.0.0
scipy==1.7.0
selenium==3.141.0
Send2Trash==1.7.1
six==1.16.0
sqlparse==0.4.1
terminado==0.10.1
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.4
vine==5.0.0
virtualenv==20.6.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
types-python-dateutil
types-requests
